+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change Log
 
+* Release 1.0.next in progress
+  * Make exception handling in shallow conversions configurable: an `:exceptions` option may be `:group`, `:omit`, `:quaify`, or `:return`.
+  * Fix bug in shallow array handling.
+
 * Release 1.0.73 on 2020-05-31
   * Add `from-java-shallow` to provide functionality similar to `clojure.core/bean` (a shallow conversion) but with options to control behavior (so "dangerous" methods that appear as getters can be omitted).
   * Bump `org.clojure/tools.logging` to `1.1.0`.